+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-#### 1.0.0b1
+#### 1.0.0b1.0.0
 
 4/25/19
 
@@ -7,3 +7,11 @@
 - add `ENABLE_PRINT` feature flag to `constants/config` and set to `false` - hides all print buttons. (Print feature is not ready.)
 - Add extended property table columns for dataset date range + counts and added `latestsaleprice` column.
 - Remove `bldgclass`, `numfloors`, and `numbldgs` columns from property tables.
+
+#### 1.0.0b1.0.1
+
+4/26/19
+
+- Add Foreclosure card to district dashboard
+- Added a post-login action to redo all authenticated data requests upon login.
+- Added post-logout redirect to main '/' to clear authenticated data from page.